fix: email-Verification stories ( add Provider)
